fix : 당첨금 총 합 오버플로우 날 가능성 배제하기 위하여 long 형으로 수정

+++ src/main/java/lotto/model/Info.java
@@ -5,11 +5,11 @@ import lotto.constant.Grade;
 
 public class Info {
     private int[] statistic;
-    private int rewards;
+    private long rewards;
 
     public Info() {
         this.statistic = new int[CommonUnits.NUM_OF_GRADES+1];
-        this.rewards = 0;
+        this.rewards = 0L;
     }
 
     public void addStatistic(Grade grade) {
@@ -21,7 +21,7 @@ public class Info {
         return statistic;
     }
 
-    public int getRewards() {
+    public long getRewards() {
         return rewards;
     }
 }
